Custom Office Templates/requirements.txt
Flask
requests
gunicorn
python-dotenv 
openai-whisper
torch
torchvision
torchaudio
huggingface_hub
pillow 